[BUGIFX] Allow editors to hide news records in list view
If an editor tries to hide a news record in list view, there is an error message
regarding access to categories.
But in list view context, the field array doesn't contain the categories,
thus the check fails. Furthermore, in list context categories are
not relevant anyway. This patch mitigates this problem.